Initial check-in of the regular expression backend.  Still pretty crappy but should work.
Solutions have to be entered as they would appear in the Java program,
i. e. you'd enter things like

  "!^.*foo.*$"

or

  "Match \\[ a bracket"

in the "Answer" field.  Invalid examples would be

  !^.*foo.*$

(missing quote marks) or

  "Match \[ a bracket"

('\[' is not valid in a Java string.  To escape a reg-exp meta char,
you need to prefix it with '\\', not just '\').

Issues:

  * Syntax check has to be improved.  In particular, we'd have to run
    the syntax check wrapper, too, to find out whether Java's
    "Pattern.compile()" raises an error.

  * Need better test environments (find all, compare sub groups, etc.)